Changing all coords in terms of PPM

+++ core/src/com/mygdx/pmd/model/components/PositionComponent.java
@@ -15,7 +15,7 @@ public class PositionComponent implements Component {
         fPos = pos;
     }
 
-    public PositionComponent(int x, int y) {
+    public PositionComponent(float x, float y) {
         this(new Vector2(x, y));
     }
